Added Save/Load functionality to the game, so we can now save (for now only scene items and grid properties) from the pause menu to a serialized file, and then load from the pause menu. Also added a check to always open the pause menu to the inventory screen, regardless of the last tab opened.
Added Save/Load functionality to the game, so we can now save (for now only scene items and grid properties) from the pause menu to a serialized file, and then load from the pause menu. Also added a check to always open the pause menu to the inventory screen, regardless of the last tab opened.