Added functionality to FULLY personalize the players skin color from the pause menu customization panel. This is done via RGB sliders, which then change all 4 skin shades on the player relative to the main skin color. It is updated real-time with a preview of the player in the pause panel
Added functionality to FULLY personalize the players skin color from the pause menu customization panel. This is done via RGB sliders, which then change all 4 skin shades on the player relative to the main skin color. It is updated real-time with a preview of the player in the pause panel