Added adornment-customization functionality to the pause menu customization panel. You can now switch between no adornments, glasses, or beard in the panel, and see it update live in the player preview sprite on the same panel. Also set up beard-tinting to the same color as the player's customized hair! Here we had to color swap the default brown beard pixels for grayscale pixels, and then tinted them to the input players hair color
Added adornment-customization functionality to the pause menu customization panel. You can now switch between no adornments, glasses, or beard in the panel, and see it update live in the player preview sprite on the same panel. Also set up beard-tinting to the same color as the player's customized hair! Here we had to color swap the default brown beard pixels for grayscale pixels, and then tinted them to the input players hair color