Added functionality to save and load character customization parameters! These are saved when you click "Save Game" from the pause menu, and then all the values are re-loaded and re-processed when you hit "Load Game". All of the toggles and sliders are also set to the loaded values when this occurs!
Added functionality to save and load character customization parameters! These are saved when you click "Save Game" from the pause menu, and then all the values are re-loaded and re-processed when you hit "Load Game". All of the toggles and sliders are also set to the loaded values when this occurs!