Added the NPC movement class to take in instructions from a NPCMovementEvent in where to go, which direction to face when they get there, and what animation to run when they get there. This occurs at a specified time, day, season, and weather pattern, with a given priority. When this scheduled event matches the current state of the game, it will build the path with the AStar algorithm, and have the NPC walk along every calculated step in that path, properly changing his/her walking direction animation, and then he/she plays an animation (i.e. smoking, digging, etc) when they get there, or idle in the correct direction.
Added the NPC movement class to take in instructions from a NPCMovementEvent in where to go, which direction to face when they get there, and what animation to run when they get there. This occurs at a specified time, day, season, and weather pattern, with a given priority. When this scheduled event matches the current state of the game, it will build the path with the AStar algorithm, and have the NPC walk along every calculated step in that path, properly changing his/her walking direction animation, and then he/she plays an animation (i.e. smoking, digging, etc) when they get there, or idle in the correct direction.