Added functionality to play ambient sounds and game music and when we are in given scenes. Each scene has it's own set of both clips, and when a scene is first loaded, this will automatically start playing ambient sounds for 20-40 seconds randomly chosen. It'll then transition into the GameMusic over 8 seconds (with soft ambient sounds in the background), and play that for 120 seconds. It'll then transition back to ambient sounds only over 8 more seconds, and play that indefinitely until we load a new scene. Also added a sound effect for when we drop an item, either by dragging or clicking.
Added functionality to play ambient sounds and game music and when we are in given scenes. Each scene has it's own set of both clips, and when a scene is first loaded, this will automatically start playing ambient sounds for 20-40 seconds randomly chosen. It'll then transition into the GameMusic over 8 seconds (with soft ambient sounds in the background), and play that for 120 seconds. It'll then transition back to ambient sounds only over 8 more seconds, and play that indefinitely until we load a new scene. Also added a sound effect for when we drop an item, either by dragging or clicking.